requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
argon2-cffi==23.1.0
pyjwt==2.8.0
python-dotenv==1.0.0

# PDF Generation
weasyprint==60.1
qrcode[pil]==7.4.2

# Templates
jinja2==3.1.2

# Scheduling
apscheduler==3.10.4

# Excel/CSV
openpyxl==3.1.2

# Image Processing
pillow==10.1.0

# Logging & Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
faker==20.1.0

# Code Quality
ruff==0.8.4

# OCR (Optional)
# pytesseract==0.3.10
# google-cloud-vision==3.4.5
# boto3==1.29.7  # for AWS Textract
